add separate bar and timeseries
